add test example
