--- example/assets/doc/test_component/test_component-library.md
@@ -1,38 +1,0 @@
-
-
-
-# test_component library
-
-
-
-
-
-
-    *[<Null safety>](https://dart.dev/null-safety)*
-
-
-
-
-
-## Classes
-
-##### [TestComponent](../test_component/TestComponent-class.md)
-
-
-
-This class is a test component 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
