+++ example/assets/lib/doc/generated_plugin_registrant/generated_plugin_registrant-library.md
@@ -1,0 +1,40 @@
+
+
+
+# generated_plugin_registrant library
+
+
+
+
+
+
+    *[<Null safety>](https://dart.dev/null-safety)*
+
+
+
+
+
+
+
+
+
+
+## Functions
+
+##### [registerPlugins](../generated_plugin_registrant/registerPlugins.md)([Registrar](https://api.flutter.dev/flutter/flutter_web_plugins/Registrar-class.html) registrar) void
+
+
+
+   
+
+
+
+
+
+
+
+
+
+
+
+
