+++ example/assets/lib/doc/list_example/ListExample-class.md
@@ -1,7 +1,7 @@
 
 
 
-# TestComponent class
+# ListExample class
 
 
 
@@ -12,7 +12,6 @@
 
 
 
-<p>This class is a test component</p>
 
 
 **Inheritance**
@@ -21,7 +20,7 @@
 - [DiagnosticableTree](https://api.flutter.dev/flutter/foundation/DiagnosticableTree-class.html)
 - [Widget](https://api.flutter.dev/flutter/widgets/Widget-class.html)
 - [StatelessWidget](https://api.flutter.dev/flutter/widgets/StatelessWidget-class.html)
-- TestComponent
+- ListExample
 
 
 
@@ -30,7 +29,7 @@
 
 ## Constructors
 
-[TestComponent](../test_component/TestComponent/TestComponent.md) ({[Key](https://api.flutter.dev/flutter/foundation/Key-class.html)? key, required [String](https://api.flutter.dev/flutter/dart-core/String-class.html) s})
+[ListExample](../list_example/ListExample/ListExample.md) ({[Key](https://api.flutter.dev/flutter/foundation/Key-class.html)? key})
 
 I got a string   _const_ 
 
@@ -64,23 +63,14 @@ _read-only, inherited_
 
 
 
-##### [s](../test_component/TestComponent/s.md) &#8594; [String](https://api.flutter.dev/flutter/dart-core/String-class.html)
-
-
-
-   
-_final_
-
-
-
 
 ## Methods
 
-##### [build](../test_component/TestComponent/build.md)([BuildContext](https://api.flutter.dev/flutter/widgets/BuildContext-class.html) context) [Widget](https://api.flutter.dev/flutter/widgets/Widget-class.html)
+##### [build](../list_example/ListExample/build.md)([BuildContext](https://api.flutter.dev/flutter/widgets/BuildContext-class.html) context) [Widget](https://api.flutter.dev/flutter/widgets/Widget-class.html)
 
 
 
-Describes the part of the user interface represented by this widget. [...](../test_component/TestComponent/build.md)  
+Describes the part of the user interface represented by this widget. [...](../list_example/ListExample/build.md)  
 _override_
