+++ example/assets/lib/doc/list_example/ListExample/ListExample.md
@@ -1,0 +1,31 @@
+
+
+
+# ListExample constructor
+
+
+
+
+    *[<Null safety>](https://dart.dev/null-safety)*
+
+
+const
+ListExample({[Key](https://api.flutter.dev/flutter/foundation/Key-class.html)? key})
+
+
+<p>I got a string</p>
+
+
+
+## Implementation
+
+```dart
+const ListExample({Key? key}) : super(key: key);
+```
+
+
+
+
+
+
+
