--- example/lib/doc/index.md
@@ -1,40 +1,0 @@
-
-
-
-# example - Dart API docs
-
-
-<h1 id="example">example</h1>
-<p>A new Flutter project.</p>
-<h2 id="getting-started">Getting Started</h2>
-<p>This project is a starting point for a Flutter application.</p>
-<p>A few resources to get you started if this is your first Flutter project:</p>
-<ul>
-<li><a href="https://flutter.dev/docs/get-started/codelab">Lab: Write your first Flutter app</a></li>
-<li><a href="https://flutter.dev/docs/cookbook">Cookbook: Useful Flutter samples</a></li>
-</ul>
-<p>For help getting started with Flutter, view our
-<a href="https://flutter.dev/docs">online documentation</a>, which offers tutorials,
-samples, guidance on mobile development, and a full API reference.</p>
-
-
-## Libraries
-
-##### [generated_plugin_registrant](generated_plugin_registrant/generated_plugin_registrant-library.md)
- 
-
-
-##### [main](main/main-library.md)
- 
-
-
-##### [test_component](test_component/test_component-library.md)
- 
-
-
-
-
-
-
-
-
