--- example/lib/doc/main/Storyboard/Storyboard.md
@@ -1,30 +1,0 @@
-
-
-
-# Storyboard constructor
-
-
-
-
-    *[<Null safety>](https://dart.dev/null-safety)*
-
-
-const
-Storyboard({[Key](https://api.flutter.dev/flutter/foundation/Key-class.html)? key})
-
-
-
-
-
-## Implementation
-
-```dart
-const Storyboard({Key? key}) : super(key: key);
-```
-
-
-
-
-
-
-
