--- example/lib/doc/main/main-library.md
@@ -1,58 +1,0 @@
-
-
-
-# main library
-
-
-
-
-
-
-    *[<Null safety>](https://dart.dev/null-safety)*
-
-
-
-
-
-## Classes
-
-##### [Storyboard](../main/Storyboard-class.md)
-
-
-
- 
-
-
-
-
-
-
-
-## Functions
-
-##### [loadAsset](../main/loadAsset.md)(dynamic context) [Future](https://api.flutter.dev/flutter/dart-async/Future-class.html)&lt;[String](https://api.flutter.dev/flutter/dart-core/String-class.html)>
-
-
-
-   
-
-
-
-
-##### [main](../main/main.md)() void
-
-
-
-   
-
-
-
-
-
-
-
-
-
-
-
-
