--- example/lib/doc/main/main.md
@@ -1,37 +1,0 @@
-
-
-
-# main function
-
-
-
-
-
-
-    *[<Null safety>](https://dart.dev/null-safety)*
-
-
-
-
-void main
-()
-
-
-
-
-
-
-## Implementation
-
-```dart
-void main() async {
-  runApp(const Storyboard());
-}
-```
-
-
-
-
-
-
-
