--- example/lib/doc/test_component/TestComponent/TestComponent.md
@@ -1,31 +1,0 @@
-
-
-
-# TestComponent constructor
-
-
-
-
-    *[<Null safety>](https://dart.dev/null-safety)*
-
-
-const
-TestComponent({[Key](https://api.flutter.dev/flutter/foundation/Key-class.html)? key, required [String](https://api.flutter.dev/flutter/dart-core/String-class.html) s})
-
-
-<p>I got a string</p>
-
-
-
-## Implementation
-
-```dart
-const TestComponent({Key? key, required this.s}) : super(key: key);
-```
-
-
-
-
-
-
-
