--- example/lib/doc/test_component/TestComponent/s.md
@@ -1,34 +1,0 @@
-
-
-
-# s property
-
-
-
-
-    *[<Null safety>](https://dart.dev/null-safety)*
-
-
-[String](https://api.flutter.dev/flutter/dart-core/String-class.html) s
-  
-_final_
-
-
-
-
-
-
-## Implementation
-
-```dart
-final String s;
-
-
-```
-
-
-
-
-
-
-
