+++ lib/src/editor/ui/doc_component_2.dart
@@ -1,26 +1,20 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_feather_icons/flutter_feather_icons.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
-import 'package:flutterbook/src/editor/ui/copy_text.dart';
-import 'package:flutterbook/src/editor/ui/styled_icon_button.dart';
 import 'package:flutterbook/src/utils/extensions.dart';
-import 'package:flutterbook/src/utils/radii.dart';
+import 'package:flutterbook/src/utils/utils.dart';
+import 'package:recase/recase.dart';
 
 class DocPanel2 extends StatefulWidget {
   final Widget component;
-  final String? docs;
-  final String? docPath;
+  final String? docName;
   final String stateName;
 
   const DocPanel2({
     Key? key,
     required this.component,
-    this.docs,
     required this.stateName,
-    this.docPath,
+    this.docName,
   }) : super(key: key);
 
   @override
@@ -28,43 +22,8 @@ class DocPanel2 extends StatefulWidget {
 }
 
 class _DocPanel2State extends State<DocPanel2> {
-  bool expanded = false;
-  double _zoom = 1;
-  get zoom => _zoom;
-
-  void zoomIn() {
-    setState(() {
-      _zoom += 0.25;
-    });
-  }
-
-  void zoomOut() {
-    setState(() {
-      _zoom = (_zoom - 0.25).clamp(0.5, 999);
-    });
-  }
-
-  void resetZoom() {
-    setState(() {
-      _zoom = 1;
-    });
-  }
-
-  void toggleExpansion() {
-    setState(() {
-      expanded = !expanded;
-    });
-  }
-
   @override
   Widget build(BuildContext context) {
-    final TextStyle tabStyle = Theme.of(context).textTheme.subtitle1!.copyWith(
-          color: context.theme.hintColor,
-          fontWeight: FontWeight.bold,
-        );
-
-    inspect(widget.component);
-    print(widget.component.toStringDeep());
     final controller = ScrollController();
     const String _markdownData = """
 ## Code blocks
@@ -80,22 +39,23 @@ Formatted Dart code looks really pretty too:
     ```
 """;
 
-    TransformationController _transformation = TransformationController();
-    _transformation.value = Matrix4.identity()..scale(zoom);
     return FutureBuilder(
-        future: loadAsset(context),
-        builder: (BuildContext context, AsyncSnapshot<String> text) {
-          return Expanded(
-            child: Markdown(
-              controller: controller,
-              data: text.data ?? _markdownData,
-            ),
-          );
-        });
+      future: loadAsset(context, widget.docName),
+      initialData: _markdownData,
+      builder: (BuildContext context, AsyncSnapshot<String> text) {
+        return Expanded(
+          child: Markdown(
+            controller: controller,
+            data: text.data ?? _markdownData,
+          ),
+        );
+      },
+    );
   }
 }
 
-Future<String> loadAsset(context) async {
-  return await rootBundle
-      .loadString('lib/doc/test_component/TestComponent/TestComponent.md');
+Future<String> loadAsset(context, docName) async {
+  ReCase doc = new ReCase(docName);
+  final docNameSnake = doc.snakeCase;
+  return await rootBundle.loadString("lib/doc/$docNameSnake/$docName/$docName.md");
 }
